docx-engine/requirements.txt
python-docx>=0.8.11
jinja2>=3.0.0
docx2pdf>=0.1.8; platform_system!='Linux'
